requirements.txt
flask==2.0.1
numpy==1.21.0
opencv-python==4.5.3.56
pytesseract==0.3.8
Pillow==8.3.1
pymongo==4.3.3
python-dotenv==0.19.2
flask-cors==2.1.3
